课程网站源码/ppt11/11.1/requirement.txt
flask
flask-sqlalchemy
mysqlclient
flask-script
flask_debugtoolbar
apscheduler
Flask-APScheduler
requests
BeautifulSoup4